refactor(api): remove unnecessary barrel exports
Remove 6 single-consumer or dead barrel files and update imports to
reference source modules directly. Keeps barrels only where they serve
a real purpose: package entry points, side-effect carriers (type
augmentations), route aggregators, or multi-file test helpers.
