--- api/packages/api/src/domain/game/index.ts
@@ -1,2 +1,0 @@
-export { encodeGameId, decodeGameId } from "./game-id.js";
-export { registerGameRoutes } from "./routes/index.js";
