--- api/packages/api/src/sources/index.ts
@@ -1,2 +1,0 @@
-export { JsonQuoteSource } from "./json-quote-source.js";
-export { StaticKeywordSource } from "./static-keyword-source.js";
