--- api/packages/game-generator/src/cipher/index.ts
@@ -1,2 +1,0 @@
-export type { GameGenerator, KeywordSource } from "./types.js";
-export { KeywordCipherGenerator } from "./keyword-cipher.js";
