--- api/packages/game-generator/src/difficulty/index.ts
@@ -1,2 +1,0 @@
-// Internal module - not exported from main package
-export { scoreDifficulty } from "./scorer.js";
