--- api/packages/game-generator/src/hints/index.ts
@@ -1,2 +1,0 @@
-// Internal module - not exported from main package
-export { generateHints } from "./generator.js";
