+++ api/packages/game-generator/src/index.ts
@@ -5,11 +5,11 @@ export type { Quote, Puzzle, Hint, CipherMapping } from "./types.js";
 export { QuoteSchema } from "./schemas.js";
 
 // Contracts
-export { QuoteSource } from "./quotes/index.js";
-export type { GameGenerator, KeywordSource } from "./cipher/index.js";
+export { QuoteSource } from "./quotes/types.js";
+export type { GameGenerator, KeywordSource } from "./cipher/types.js";
 
 // Implementations
-export { KeywordCipherGenerator } from "./cipher/index.js";
+export { KeywordCipherGenerator } from "./cipher/keyword-cipher.js";
 
 // Data
 export { KEYWORDS } from "./data/keywords.js";
