--- api/packages/game-generator/src/quotes/index.ts
@@ -1,1 +1,0 @@
-export { QuoteSource } from "./types.js";
